HistoryAdapter refactoring

+++ data/src/main/java/kg/nurtelecom/data/history_by_id/ReceiptItems.kt
@@ -6,7 +6,7 @@ data class ReceiptItems (
 	val productName : String,
 	val productCode : String,
 	val productQuantity : Int,
-	val productUnitPrice : Int,
+	val productUnitPrice : Double,
 	val productUnitMeasure : String,
 	val productNdsType : Int,
 	val productNdsAmount : Double,
